두자리수의 연산 시 Error 문자를 Insert 할때 ex) 10이면 1.0 0.0 이런식으로 붙여짐 결과값 1.00.0

+++ app/src/main/java/com/example/mobile_programming_term_project/makeCalculationStack.java
@@ -3,7 +3,7 @@ package com.example.mobile_programming_term_project;
 public interface makeCalculationStack {
     boolean isEmpty();
     boolean isFull();
-    void push(float item);
+    void push(int item);
     float pop();
     float peek();
 }
